tests: tests for new commands
